Make IlValue into a more extensible class
In preparation for creating a JitBuilder extension for IlValue,
to facilitate the client API creation, the IlValue class needs
to become more like a fully extensible with an independent header.
