+++ compiler/ilgen/CMakeLists.txt
@@ -22,7 +22,7 @@
 compiler_library(ilgen
 	${CMAKE_CURRENT_LIST_DIR}/IlGenRequest.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRIlBuilder.cpp
-	${CMAKE_CURRENT_LIST_DIR}/IlValue.cpp
+	${CMAKE_CURRENT_LIST_DIR}/OMRIlValue.cpp
 	${CMAKE_CURRENT_LIST_DIR}/IlInjector.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRMethodBuilder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRBytecodeBuilder.cpp
